fix(python): move verify_request_count to conftest.py in wire tests
