+++ seed/python-sdk/exhaustive/no-custom-config/tests/wire/test_reqWithHeaders.py
@@ -1,34 +1,7 @@
-from typing import Optional, Dict, Any
 from seed import SeedExhaustive
 
 import pytest
 
-import requests
-
-
-
-def verify_request_count(
-    test_id: str,
-    method: str,
-    url_path: str,
-    query_params: Optional[Dict[str, str]],
-    expected: int,
-) -> None:
-    """Verifies the number of requests made to WireMock filtered by test ID for concurrency safety"""
-    wiremock_admin_url = "http://localhost:8080/__admin"
-    request_body: Dict[str, Any] = {
-            "method": method,
-            "urlPath": url_path,
-            "headers": {"X-Test-Id": {"equalTo": test_id}}
-        }
-    if query_params:
-            query_parameters = {k: {"equalTo": v} for k, v in query_params.items()}
-            request_body["queryParameters"] = query_parameters
-    response = requests.post(f"{wiremock_admin_url}/requests/find", json=request_body)
-    assert response.status_code == 200, "Failed to query WireMock requests"
-    result = response.json()
-    requests_found = len(result.get("requests", []))
-    assert requests_found == expected, f"Expected {expected} requests, found {requests_found}"
 
 
 def test_reqWithHeaders_get_with_custom_header() -> None:
